+++ examples/todomvc/Cargo.toml
@@ -4,6 +4,7 @@ members = [
     "backend",
     "shared",
 ]
+resolver = "2"
 
 [workspace.package]
 version = "0.1.0"
